chore: moved directory delete files
