+++ src/main/java/com/example/webscraper/meal/controller/MealItemController.java
@@ -1,23 +1,22 @@
 package com.example.webscraper.meal.controller;
 
 import com.example.webscraper.meal.entity.MealItem;
+import com.example.webscraper.meal.service.MealItemService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.example.webscraper.meal.service.WebScraperService;
 
 @RestController
 @RequestMapping("/api/team04/scraper")
 @RequiredArgsConstructor
-public class WebScraperController {
+public class MealItemController {
 
-    private final WebScraperService webScraperService;
+    private final MealItemService mealItemService;
 
     @GetMapping("/meal")
     public ResponseEntity<MealItem> getMeal() {
-        return ResponseEntity.ok(
-            webScraperService.getMeal());
+        return ResponseEntity.ok();
     }
 }
