+++ src/main/java/com/example/webscraper/meal/entity/MealItem.java
@@ -1,15 +1,24 @@
 package com.example.webscraper.meal.entity;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.jsoup.nodes.Element;
+import webscraper.meal.model.MenuItem;
 
 @Entity
 @Builder
@@ -22,11 +31,68 @@ public class MealItem {
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private long id;
 
+  private String day;
   private String mealType;
   private String menuTitle;
   private String menuContent;
   private String extraInfo;
 
-  @Builder.Default
-  private LocalDateTime day = LocalDateTime.now();
+  public static MenuItem of(String day, Element mealType, Element title, Element content, Element extra) {
+    return new MenuItem(
+        day,
+        mealType.text().trim(),
+        title.text().trim(),
+        content.text().trim(),
+        extra.text().trim()
+    );
+  }
+
+  public String getFormattedDate() {
+    String datePart = day.split(" ")[0];
+    String[] parts = datePart.split("\\.");
+    int year = LocalDate.now().getYear();
+    return String.format("%d-%s-%s", year, parts[0], parts[1]);
+  }
+
+  public String generateId() {
+    String digit = switch (mealType) {
+      case "조식" -> "1";
+      case "중식" -> "2";
+      case "석식" -> "3";
+      default -> "0";
+    };
+    return getFormattedDate().replaceAll("-", "") + digit;
+  }
+
+  public String getFullContents() {
+    return String.join(" ", menuTitle, menuContent, extraInfo).trim();
+  }
+
+
+  public String toMealJson() {
+    ObjectMapper mapper = new ObjectMapper();
+    Map<String, Object> jsonMap = new HashMap<>();
+
+    jsonMap.put("id", Long.parseLong(generateId()));
+    jsonMap.put("dayInfo", getFormattedDate());
+    jsonMap.put("mealType", switch (mealType) {
+      case "조식" -> "BREAKFAST";
+      case "중식" -> "LUNCH";
+      case "석식" -> "DINNER";
+      default -> "UNKNOWN";
+    });
+
+    List<String> menuNames = new ArrayList<>();
+    if (!menuTitle.isBlank()) menuNames.add(menuTitle);
+    if (!menuContent.isBlank()) menuNames.add(menuContent);
+    if (!extraInfo.isBlank()) menuNames.add(extraInfo);
+
+    jsonMap.put("menuNames", menuNames);
+
+    try {
+      return mapper.writeValueAsString(jsonMap);
+    } catch (JsonProcessingException e) {
+      return "{}";
+    }
+  }
 }
