+++ src/main/java/com/example/webscraper/meal/service/MenuScraper.java
@@ -1,4 +1,4 @@
-package webscraper.meal.service;
+package com.example.webscraper.meal.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.jsoup.Jsoup;
